+++ CHANGELOG.md
@@ -12,8 +12,14 @@
     based endpoints. Can also be used to set headers for all requests.
   - Added support for Depot cloud-based caching: https://depot.dev/docs/cache/overview
   - Added support for the HTTP protocol: https://bazel.build/remote/caching#http-caching
+  - Added support for the ByteStream API.
+  - Added timeout and concurrency limit to clients.
   - Added a `MOON_DEBUG_REMOTE` environment variable, which can be used to debug internal errors for
     diagnosing connection/integration issues.
+
+#### 🐞 Fixes
+
+- Fixed lots of issues around remote caching.
 
 ## 1.31.3